requirements.txt
aiogram>=3.4,<4
SQLAlchemy[asyncio]>=2.0,<3
aiosqlite>=0.19
asyncpg>=0.29
python-dotenv>=1.0
greenlet>=3.0.3
